fix:病院情報の取得、更新を修正
