+++ src/app/Application/Service/Auth/AuthActorService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Application\Service;
+namespace App\Application\Service\Auth;
 
 use App\Exceptions\UnauthorizedException;
 use App\Models\StaffModel;
