+++ src/app/Application/Service/Hospital/Menu/GetMenuDetailService.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace App\Application\Service\Menu;
+namespace App\Application\Service\Hospital\Menu;
 
-use App\Application\Service\AuthActorService;
+use App\Application\Service\Auth\AuthActorService;
 use App\Domains\Menu\Repository\MenuRepositoryInterface;
 use App\Models\Menu;
 
-class CreateMenuService
+class GetMenuDetailService
 {
     public function __construct(
         private AuthActorService $authActorService,
@@ -16,16 +16,12 @@ class CreateMenuService
     ) {
     }
 
-    public function execute(string $name, string $detail, int $requiredTime, bool $isPublished): Menu
+    public function execute(int $id): Menu
     {
         $hospitalId = $this->authActorService->getHospitalId();
-
-        return $this->menuRepository->create(
+        return        $this->menuRepository->getByIdAndHospital(
             hospitalId: $hospitalId,
-            name: $name,
-            detail: $detail,
-            requiredTime: $requiredTime,
-            isPublished: $isPublished,
+            id: $id,
         );
     }
 }
