+++ src/app/Application/Service/ReviewService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Application\Service;
 
+use App\Application\Service\Auth\AuthActorService;
 use App\Domains\Hospital\Repositories\HospitalRepositoryInterface;
 use App\Domains\Review\Enum\Rating;
 use App\Domains\Review\Repository\ReviewRepositoryInterface;
